refactor(saga): separate definition and execution internals
