+++ .changeset/neat-sagas-organize.md
@@ -1,0 +1,5 @@
+---
+"@redemeine/saga": patch
+---
+
+Separate saga definition token internals from execution handler internals while preserving root exports and legacy internal import paths.
